Move ScopedTojos class out of SafeMojo
